Move some contracting stuff to the a separate utility module
